moved startup checks to launch_common, python venv mandatory due to Ubuntu > 23.xx, cleaned up checks for better understanding
