+++ readme.md
@@ -39,7 +39,7 @@ The [terms of use](https://api.tzkt.io/#section/Terms-of-Use) of TzKT API allow 
 
 ## Requirements and Setup
 
-Python 3 is required. You can use the following commands to install it.
+The setup instructions are Linux specific. Python 3 is required. You can use the following commands to install it.
 
 ```bash
 sudo apt-get update
@@ -52,20 +52,24 @@ Download the application repository using git clone:
 git clone https://github.com/tezos-reward-distributor-organization/tezos-reward-distributor
 ```
 
-To install required modules, use pip with `requirements.txt` provided.
+To install required modules, use pip with `requirements.txt` provided. Follow the instructions to create a virtual environment for your project specific python installation: https://packaging.python.org/en/latest/guides/installing-using-pip-and-virtual-environments/
+
+```bash
+python3 -m venv env
+source env/bin/activate
+```
 
 ```bash
 cd tezos-reward-distributor
-pip install -r requirements.txt
+python3 -m pip install -r requirements.txt
 ```
 
 To install the required modules for developers, use pip with `requirements_developer.txt` provided.
 
 ```bash
 cd tezos-reward-distributor
-pip install -r requirements_developer.txt
+python3 -m pip install -r requirements_developer.txt
 ```
-
 
 Regularly check and upgrade to the latest available version:
