+++ src/util/parser.py
@@ -1,5 +1,6 @@
-import argparse
 import os
+import argparse
+from util.args_validator import validate
 
 from Constants import (
     BASE_DIR,
@@ -268,3 +269,10 @@ def add_argument_log_file(argparser):
         ),
         default=default_log_file,
     )
+
+
+def parse_arguments(args=None):
+    parser = build_parser()
+    # Basic validations
+    args = validate(parser)
+    return args
